Extract some logic from index.ts into helpers.ts (pure refactoring)
